Move `deep_flatten` to utils.py
